+++ src/redturtle/prenotazioni/tests/test_prenotazione_events.py
@@ -77,7 +77,7 @@ class TestSPrenotazioneEvents(unittest.TestCase):
         booker = IBooker(self.folder_prenotazioni)
         if booking_date is None:
             booking_date = self.tomorrow_8_0
-        return booker.create(
+        return booker.book(
             {
                 "booking_date": booking_date,
                 "booking_type": "Type A",