+++ src/redturtle/prenotazioni/tests/test_prenotazioni_context_state.py
@@ -2,14 +2,13 @@
 import unittest
 from datetime import date
 from datetime import datetime
+from datetime import timedelta
 
 from Acquisition import aq_parent
 from plone import api
-from plone.app.testing import SITE_OWNER_NAME
-from plone.app.testing import SITE_OWNER_PASSWORD
+from plone.app.testing import logout, login
 from plone.app.testing import TEST_USER_ID
 from plone.app.testing import setRoles
-from plone.restapi.testing import RelativeSession
 
 from redturtle.prenotazioni.adapters.booker import IBooker
 from redturtle.prenotazioni.testing import REDTURTLE_PRENOTAZIONI_INTEGRATION_TESTING
@@ -26,17 +25,21 @@ class TestPrenotazioniContextState(unittest.TestCase):
         self.portal_url = self.portal.absolute_url()
         setRoles(self.portal, TEST_USER_ID, ["Manager"])
 
-        self.api_session_admin = RelativeSession(self.portal_url)
-        self.api_session_admin.headers.update({"Accept": "application/json"})
-        self.api_session_admin.auth = (SITE_OWNER_NAME, SITE_OWNER_PASSWORD)
+        api.user.create(
+            email="user@example.com",
+            username="jdoe",
+            password="secret!!!",
+        )
 
-        self.portal_url = self.portal.absolute_url()
+        api.user.grant_roles(username="jdoe", roles=["Bookings Manager"])
+
+        self.today = date.today()
         self.folder_prenotazioni = api.content.create(
             container=self.portal,
             type="PrenotazioniFolder",
             title="Folder",
             description="",
-            daData=date.today(),
+            daData=self.today,
             gates=["Gate A"],
             week_table=WEEK_TABLE_SCHEMA,
         )
@@ -48,6 +51,13 @@ class TestPrenotazioniContextState(unittest.TestCase):
             container=self.folder_prenotazioni,
             gates=["all"],
         )
+        api.content.transition(obj=self.folder_prenotazioni, transition="publish")
+
+        self.prenotazioni_context_state = api.content.get_view(
+            context=self.folder_prenotazioni,
+            request=self.request,
+            name="prenotazioni_context_state",
+        )
 
     def test_get_free_slots_skip_bookigs_inside_pause_range(self):
         booker = IBooker(self.folder_prenotazioni)
@@ -55,7 +65,7 @@ class TestPrenotazioniContextState(unittest.TestCase):
         today = date.today()
         # need this just to have the day container
         aq_parent(
-            booker.create(
+            booker.book(
                 {
                     "booking_date": datetime(
                         today.year, today.month, today.day, 10, 00
@@ -77,18 +87,25 @@ class TestPrenotazioniContextState(unittest.TestCase):
             {"day": "7", "pause_end": "1100", "pause_start": "0800"},
         ]
 
-        view = api.content.get_view(
-            context=self.folder_prenotazioni,
-            request=self.request,
-            name="prenotazioni_context_state",
-        )
-        res = view.get_free_slots(today)
+        res = self.prenotazioni_context_state.get_free_slots(today)
         # available slots are only arount the pause and not inside it
         self.assertEqual(len(res["Gate A"]), 2)
-        self.assertEqual(view.get_free_slots(today)["Gate A"][0].start(), "07:00")
-        self.assertEqual(view.get_free_slots(today)["Gate A"][0].stop(), "08:00")
-        self.assertEqual(view.get_free_slots(today)["Gate A"][1].start(), "11:00")
-        self.assertEqual(view.get_free_slots(today)["Gate A"][1].stop(), "13:00")
+        self.assertEqual(
+            self.prenotazioni_context_state.get_free_slots(today)["Gate A"][0].start(),
+            "07:00",
+        )
+        self.assertEqual(
+            self.prenotazioni_context_state.get_free_slots(today)["Gate A"][0].stop(),
+            "08:00",
+        )
+        self.assertEqual(
+            self.prenotazioni_context_state.get_free_slots(today)["Gate A"][1].start(),
+            "11:00",
+        )
+        self.assertEqual(
+            self.prenotazioni_context_state.get_free_slots(today)["Gate A"][1].stop(),
+            "13:00",
+        )
 
     def test_get_free_slots_handle_pauses_correctly(self):
         booker = IBooker(self.folder_prenotazioni)
@@ -96,7 +113,7 @@ class TestPrenotazioniContextState(unittest.TestCase):
         today = date.today()
         # need this just to have the day container
         aq_parent(
-            booker.create(
+            booker.book(
                 {
                     "booking_date": datetime(
                         today.year, today.month, today.day, 10, 30
@@ -108,7 +125,7 @@ class TestPrenotazioniContextState(unittest.TestCase):
         )
         for hour in [7, 8, 9, 11, 12]:
             aq_parent(
-                booker.create(
+                booker.book(
                     {
                         "booking_date": datetime(
                             today.year, today.month, today.day, hour, 00
@@ -120,7 +137,7 @@ class TestPrenotazioniContextState(unittest.TestCase):
             )
 
             aq_parent(
-                booker.create(
+                booker.book(
                     {
                         "booking_date": datetime(
                             today.year, today.month, today.day, hour, 30
@@ -142,11 +159,36 @@ class TestPrenotazioniContextState(unittest.TestCase):
             {"day": "7", "pause_end": "1030", "pause_start": "1000"},
         ]
 
-        view = api.content.get_view(
-            context=self.folder_prenotazioni,
-            request=self.request,
-            name="prenotazioni_context_state",
-        )
-        res = view.get_free_slots(today)
+        res = self.prenotazioni_context_state.get_free_slots(today)
         # there are no free slots because are all taken by bookings or pause
         self.assertEqual(len(res["Gate A"]), 0)
+
+    def test_if_futureDays_is_not_set_maximum_bookable_date_return_none(self):
+        logout()
+
+        self.assertIsNone(self.prenotazioni_context_state.maximum_bookable_date)
+
+    def test_if_futureDays_is_0_maximum_bookable_date_return_none(self):
+        self.folder_prenotazioni.futureDays = 0
+        logout()
+        self.assertIsNone(self.prenotazioni_context_state.maximum_bookable_date)
+
+    def test_if_futureDays_is_set_maximum_bookable_date_return_right_date_for_normal_users(
+        self,
+    ):
+        self.folder_prenotazioni.futureDays = 2
+        logout()
+        self.assertEqual(
+            self.prenotazioni_context_state.maximum_bookable_date.date(),
+            self.today + timedelta(days=2),
+        )
+
+    def test_managers_bypass_futureDays_in_maximum_bookable_date(
+        self,
+    ):
+        self.folder_prenotazioni.futureDays = 2
+        self.assertIsNone(self.prenotazioni_context_state.maximum_bookable_date)
+
+        logout()
+        login(self.portal, "jdoe")
+        self.assertIsNone(self.prenotazioni_context_state.maximum_bookable_date)